Exclude asadm directory
Prior to aerospike/aerospike-tools:7.0.5, /opt/aerospike/bin/asadm was a binary bundle that was automatically excluded from copying by this script. But starting in 7.0.5, it is now an unpacked directory with .so and .py files. We want to exclude the entire bin/asadm directory because it is already available at https://github.com/aerospike/aerospike-admin.